Created firebase wrapper
